+++ src/fraster_extract/__init__.py
@@ -1,1 +1,1 @@
-from .extract import *
+from ._stats import *
